# [Android] Optimize Gradle build