--- a/android/build.gradle
+++ b/android/build.gradle
@@ -186,16 +186,18 @@ android {
     }
 
     if (isGHExampleApp()) {
-        tasks.withType(ExternalNativeBuildJsonTask) {
-            compileTask ->
-                compileTask.doLast {
-                    def rootDir = new File("${project.projectDir}/..")
-                    def generated = new File("${compileTask.abi.getCxxBuildFolder()}/compile_commands.json")
-                    def output = new File("${rootDir}/compile_commands.json")
-                    output.text = generated.text
+        tasks.withType(ExternalNativeBuildJsonTask).tap {
+            configureEach {
+                compileTask ->
+                    compileTask.doLast {
+                        def rootDir = new File("${project.projectDir}/..")
+                        def generated = new File("${compileTask.abi.getCxxBuildFolder()}/compile_commands.json")
+                        def output = new File("${rootDir}/compile_commands.json")
+                        output.text = generated.text
 
-                    println("Generated clangd metadata.")
-                }
+                        println("Generated clangd metadata.")
+                    }
+            }
         }
     }
 }