## Description

Results for nth build:

baseline:
- Total build time 6.776s
- Total build time 7.391s ([link](https://scans.gradle.com/s/6sz4ssuepvema))
optimised:
- Total build time 0.699s ([link](https://gradle.com/s/bzogua7a3ki6g))
- Total build time 0.827s ([link](https://gradle.com/s/ogpadox4x2qls))

fixes: #2865

## Test plan

- Run `./gradlew help --scan` to see pre-build performance
- Build app to see everything still works